refactor: 홈에 쓰이는 이미지 에셋 png -> webp로 변환 적용 (#27)
